% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epic_unmix_func.R
\name{run_epic_unmix}
\alias{run_epic_unmix}
\title{EPIC-unmix for sample-level cell-type-specific gene expression analysis}
\usage{
run_epic_unmix(
  bulk,
  frac,
  input_cts,
  outf = TRUE,
  out_prefix = NULL,
  out_ct = NULL,
  nstop = 1,
  delta = 0.1,
  nu0 = 50,
  nu1 = 50,
  seed = 1,
  ncore = NULL
)
}
\arguments{
\item{bulk}{bulk gene expression (gene x sample). We recommend log2-transformed data for better performance. If the max(bulk) > 50, bulk will be transformed to log2(count per million + 1).}

\item{frac}{sample-specific cell type fraction (sample x cell type). This could be estimated by external software, e.g., MuSiC, bMIND, etc.}

\item{input_cts}{an R object, standard output from `bMIND` or `bMIND_de` in this package. It will be used to infer the second layer of prior.}

\item{outf}{logical, write output (CTS profile) to files or not, one file for each cell type. Default TRUE}

\item{out_prefix}{file output name prefix. Default value "epic_unmix"}

\item{out_ct}{selected output cell types. Default all the cell types}

\item{nstop}{number of iteration before early stop. In practice, we do not recommend iterating too many times, since it will significantly reduce the number of genes 
(it will drop genes during each iteration). In our evaluations, 1 time iteration would already give satisfying performance.}

\item{delta}{stoping criteria. The maximum mean difference between two iterations. Default 0.1. We do not recommend setting this to be very stringent, 
which will significantly increase the number of iterations needed and will drop genes.}

\item{nu0}{hyper-parameter for the prior covariance matrix initially derived from the reference. The larger the nu0, the higher the certainty about the information in covariance, and the more 
informative is the distribution. The default is 50.}

\item{nu1}{hyper-parameter for the prior covariance matrix derived from the last iteration. The larger the nu1, the more informative after each iteration. Default is 50.}

\item{seed}{random seed used. Default 1.}

\item{ncore}{number of cores to run in parallel. The default is all available cores.}
}
\value{
Same as bMIND output, a list containing of the following elements:
\item{A}{the deconvolved cell-type-specific gene expression (gene x cell type x sample). Note that if outf == TRUE, it will be written to files.}
\item{SE}{the standard error of cell-type-specific gene expression (gene x cell type x sample).}
\item{Sigma_c}{the covariance matrix for the deconvolved cell-type-specific expression (gene x cell type x cell type).}
\item{mu}{the estimated profile matrix (gene x cell type).}
\item{frac}{the estimated cell type fractions (sample x cell type) if fractions are not provided.}
}
\description{
A two step Bayesian method. In the first step, a normal Bayesian framework was adopted, similar to the bMIND algorithm. In the second step, we update the prior using results from the first step.
}
\examples{

file_dir <- 'example/'
bulk <- read.table(paste0(file_dir,'by_id.csv'),sep=',',header=T, row.name=1,check.names=F)
frac <- read.table(paste0(file_dir,'fracs.csv'),sep = ',',header=T,check.names=F)
cts = readRDS("bmind.rds")
#note that the example already matched genes, otherwise genes must be exactly matched
epic_unmix = run_epic_unmix(bulk, frac, cts, out_prefix = paste0(file_dir,'epic_unmix'))

}
